feat: 모듈화와 렌더 엔트리 파일 main.tsx 생성
