+++ src/index.html
@@ -1,2 +1,2 @@
 <div id="myapp"></div>
-<script type="module" src="../dist/app.js"></script>
+<script type="module" src="../dist/main.js"></script>
